Rename shapefile-ingest to analytic-data
