+++ analytic-data/common/src/main/scala/com/azavea/ingest/common/CSVSchemaParser.scala
