+++ analytic-data/geomesa/src/main/scala/com/azavea/ingest/geomesa/Main.scala
