+++ analytic-data/project/Version.scala
