+++ analytic-data/project/build.properties
